+++ src/handle_macho_file_parse_result.c
@@ -783,7 +783,7 @@ handle_macho_file_parse_result(struct tbd_for_main *const global,
         }
     }
 
-    const uint64_t flags_re = tbd->flags_re;
+    const uint32_t flags_re = tbd->flags_re;
     if (flags_re != 0) {
         if (tbd->options & O_TBD_FOR_MAIN_ADD_OR_REMOVE_FLAGS) {
             tbd->info.flags_field &= ~flags_re;
